Move Notation to a separate utility file (in response to Volker's review).
