+++ mathjax3-ts/output/chtml/Wrappers/menclose.ts
@@ -25,6 +25,7 @@ import {CHTMLWrapper, StringMap} from '../Wrapper.js';
 import {CHTMLWrapperFactory} from '../WrapperFactory.js';
 import {CHTMLmsqrt} from './msqrt.js';
 import {BBox} from '../BBox.js';
+import * as Notation from '../Notation.js';
 import {MmlMenclose} from '../../../core/MmlTree/MmlNodes/menclose.js';
 import {MmlNode, AbstractMmlNode, AttributeList} from '../../../core/MmlTree/MmlNode.js';
 import {OptionList} from '../../../util/Options.js';
@@ -33,296 +34,6 @@ import {split} from '../../../util/string.js';
 
 /*****************************************************************/
 
-const ARROWX = 4, ARROWDX = 1, ARROWY = 2;  // default relative arrowhead values
-
-const THICKNESS = .067;  // default rule thickness
-const PADDING = .2;      // default padding
-
-const SOLID = THICKNESS + 'em solid';  // a solid border
-
-/*
- * The functions used for notation definitions
- */
-export type Renderer<N, T, D> = (node: CHTMLmenclose<N, T, D>, child: N) => void;
-export type BBoxExtender<N, T, D> = (node: CHTMLmenclose<N, T, D>) => number[];
-export type BBoxBorder<N, T, D> = (node: CHTMLmenclose<N, T, D>) => number[];
-export type Initializer<N, T, D> = (node: CHTMLmenclose<N, T, D>) => void;
-
-/*
- * The definition of a notation
- */
-export type NotationDef<N, T, D> = {
-    renderer: Renderer<N, T, D>;        // renders the HTML for the notation
-    bbox: BBoxExtender<N, T, D>;        // gives the offsets to the child bounidng box: [top, right, bottom, left]
-    border?: BBoxBorder<N, T, D>;       // gives the amount of the bbox offset that is due to borders on the child
-    renderChild?: boolean;              // true if the notation is used to render the child directly (e.g., radical)
-    init?: Initializer<N, T, D>;        // function to be called during wrapper construction
-    remove?: string;                    // list of notations that are suppressed by this one
-};
-
-/*
- * For defining notation maps
- */
-export type NotationDefPair = [string, NotationDef<any, any, any>];
-export type NotationDefList = Map<string, NotationDef<any, any, any>>;
-
-/*
- * The list of notations for an menclose element
- */
-export type NotationList<N, T, D> = {[notation: string]: NotationDef<N, T, D>};
-
-
-/*****************************************************************/
-
-/*
- * The names and indices of sides for borders, padding, etc.
- */
-export const sideIndex = {top: 0, right: 1, bottom: 2, left: 3};
-export type Side = keyof typeof sideIndex;
-
-/*
- * the spacing to leave for an arrowhead
- */
-export const arrowHead = (node: CHTMLmenclose<any, any, any>) => {
-    return Math.max(node.padding, node.thickness * (node.arrowhead.x + node.arrowhead.dx + 1));
-};
-
-/*
- * Service routines used for notations
- */
-export const Notation = {
-
-    /*
-     * Common BBox and Border functions
-     */
-    fullBBox: ((node) => new Array(4).fill(node.thickness + node.padding)) as BBoxExtender<any, any, any>,
-    fullPadding: ((node) => new Array(4).fill(node.padding)) as BBoxExtender<any, any, any>,
-    fullBorder: ((node) => new Array(4).fill(node.thickness)) as BBoxBorder<any, any, any>,
-
-    /*
-     * The data for horizontal and vertical arrow notations
-     *   [angle, neg, double, origin, offset, isVertical, remove]
-     */
-    arrowDef: {
-        up:    [-Math.PI / 2,  1, false, 'bottom left', 'X', true,  'verticalstrike'],
-        down:  [ Math.PI / 2, -1, false, 'top left',    'X', true,  'verticakstrike'],
-        right: [  0,          -1, false, 'top left',    'Y', false, 'horizontalstrike'],
-        left:  [  Math.PI,    -1, false, 'top right',   'Y', false, 'horizontalstrike'],
-        updown:    [Math.PI / 2, -1, true, 'top left',  'X', true,  'verticalstrike uparrow downarrow'],
-        leftright: [0,           -1, true, 'top left',  'Y', false, 'horizontalstrike leftarrow rightarrow']
-    } as {[name: string]: [number, number, boolean, string, string, boolean, string]},
-
-    /*
-     * The data for diagonal arrow notations
-     *   [neg, c, pi, double, origin, remove]
-     */
-    diagonalArrowDef: {
-        updiagonal: [ 1, -1, 0,       false, 'bottom left',  'updiagonalstrike'],
-        northeast:  [ 1, -1, 0,       false, 'bottom left',  'updiagonalstrike'],
-        southeast:  [-1,  1, 0,       false, 'top left',     'downdiagonalstrike'],
-        northwest:  [ 1,  1, Math.PI, false, 'bottom right', 'downdiagonalstrike'],
-        southwest:  [-1, -1, Math.PI, false, 'top right',    'updiagonalstrike'],
-        northeastsouthwest: [ 1, -1, 0, true, 'bottom left', 'updiagonalstrike northeastarrow updiagonalarrow southwestarrow'],
-        northwestsoutheast: [-1,  1, 0, true, 'top left',    'downdiagonalstrike northwestarrow southeastarrow']
-    } as {[name: string]: [number, number, number, boolean, string, string]},
-
-    /*
-     * The BBox functions for horizontal and vertical arrows
-     */
-    arrowBBox: {
-        up:    (node) => [arrowHead(node), 0, node.padding, 0],
-        down:  (node) => [node.padding, 0, arrowHead(node), 0],
-        right: (node) => [0, arrowHead(node), 0, node.padding],
-        left:  (node) => [0, node.padding, 0, arrowHead(node)],
-        updown:    (node) => [arrowHead(node), 0, arrowHead(node), 0],
-        leftright: (node) => [0, arrowHead(node), 0, arrowHead(node)]
-    } as {[name: string]: BBoxExtender<any, any, any>},
-
-    /*
-     * Create a named element (handled by CSS), and adjust it if thickness is non-standard
-     *
-     * @param{string} name    The name of the element to create
-     * @param{string} offset  The offset direction to adjust if thickness is non-standard
-     * @return{Renderer}      The renderer function for the given element name
-     */
-    RenderElement: (name: string, offset: string = '') => {
-        return ((node, child) => {
-            const shape = node.adjustBorder(node.html('mjx-' + name));
-            if (offset && node.thickness !== THICKNESS) {
-                const transform = 'translate' + offset + '(' + node.em(node.thickness / 2) + ')';
-                node.adaptor.setStyle(shape, 'transform', transform);
-            }
-            node.adaptor.append(node.chtml, shape);
-        }) as Renderer<any, any, any>;
-    },
-
-    /*
-     * @param{string} side       The side on which a border should appear
-     * @return{NotationDefPair}  The notation definition for the notation having a line on the given side
-     */
-    Border: (side: Side) => {
-        const i = sideIndex[side];
-        return [side, {
-            //
-            // Add the border to the main child object
-            //
-            renderer: (node, child) => {
-                node.adaptor.setStyle(child, 'border-' + side, node.em(node.thickness) + ' solid');
-            },
-            //
-            // Indicate the extra space on the given side
-            //
-            bbox: (node) => {
-                const bbox = [0, 0, 0, 0];
-                bbox[i] = node.thickness + node.padding;
-                return bbox;
-            },
-            //
-            // Indicate the border on the given side
-            //
-            border: (node) => {
-                const bbox = [0, 0, 0, 0];
-                bbox[i] = node.thickness;
-                return bbox;
-            }
-        }] as NotationDefPair;
-    },
-
-    /*
-     * @param{string} name       The name of the notation to define
-     * @param{string} side1      The first side to get a border
-     * @param{string} side2      The second side to get a border
-     * @return{NotationDefPair}  The notation definition for the notation having lines on two sides
-     */
-    Border2: (name: string, side1: Side, side2: Side) => {
-        const i1 = sideIndex[side1];
-        const i2 = sideIndex[side2];
-        return [name, {
-            //
-            // Add the border along the given sides
-            //
-            renderer: (node, child) => {
-                const border = node.em(node.thickness) + ' solid';
-                node.adaptor.setStyle(child, 'border-' + side1, border);
-                node.adaptor.setStyle(child, 'border-' + side2, border);
-            },
-            //
-            // Mark the extra space along the two sides
-            //
-            bbox: (node) => {
-                const t = node.thickness + node.padding;
-                const bbox = [0, 0, 0, 0];
-                bbox[i1] = bbox[i2] = t;
-                return bbox;
-            },
-            //
-            // Indicate the border on the two sides
-            //
-            border: (node) => {
-                const bbox = [0, 0, 0, 0];
-                bbox[i1] = bbox[i2] = node.thickness;
-                return bbox;
-            },
-            //
-            // Remove the single side notations, if present
-            //
-            remove: side1 + ' ' + side2
-        }] as NotationDefPair;
-    },
-
-    /*
-     * @param{string} name       The name of the diagonal strike to define
-     * @param{number} neg        1 or -1 to use with the angle
-     * @return{NotationDefPair}  The notation definition for the diagonal strike
-     */
-    DiagonalStrike: (name: string, neg: number) => {
-        const cname = 'mjx-' + name.charAt(0) + 'strike';
-        return [name + 'diagonalstrike', {
-            //
-            // Find the angle and width from the bounding box size and create
-            //   the diagonal line from them by rotating a border of the proper size
-            //
-            renderer: (node, child) => {
-                const {w, h, d} = node.getBBox();
-                const [a, W] = node.getArgMod(w, h + d);
-                const t = neg * node.thickness / 2;
-                const strike = node.adjustBorder(node.html(cname, {style: {
-                    width: node.em(W),
-                    transform: 'rotate(' + node.units(-neg * a) + 'rad) translateY(' + t + 'em)',
-                }}));
-                node.adaptor.append(node.chtml, strike);
-            },
-            //
-            //  Add padding all around
-            //
-            bbox: Notation.fullBBox
-        }] as NotationDefPair;
-    },
-
-    /*
-     * @param{string} name       The name of the diagonal arrow to define
-     * @return{NotationDefPair}  The notation definition for the diagonal arrow
-     */
-    DiagonalArrow: (name: string) => {
-        const [neg, c, pi, double, origin, remove] = Notation.diagonalArrowDef[name];
-        return [name + 'arrow', {
-            //
-            // Find the angle and width from the bounding box size and create
-            //   the arrow from them and the other arrow data
-            //
-            renderer: (node, child) => {
-                const {w, h, d} = node.getBBox();
-                const [a, W] = node.getArgMod(w, h + d);
-                const arrow = node.arrow(W, c * (a - pi), neg, origin, '', 0, double);
-                node.adaptor.append(node.chtml, arrow);
-            },
-            //
-            // Add roughly the right amount of space for the arrowhead all around
-            //
-            bbox: (node) => {
-                const t = Math.max(node.padding, node.thickness * (node.arrowhead.x + node.arrowhead.dx) / Math.sqrt(2));
-                return [t, t, t, t];
-            },
-            //
-            // Remove redundant notations
-            //
-            remove: remove
-        }] as NotationDefPair;
-    },
-
-    /*
-     * @param{string} name       The name of the horizontal or vertical arrow to define
-     * @return{NotationDefPair}  The notation definition for the arrow
-     */
-    Arrow: (name: string) => {
-        const [angle, neg, double, origin, offset, isVertical, remove] = Notation.arrowDef[name];
-        return [name + 'arrow', {
-            //
-            // Get the arrow height and depth from the bounding box and the arrow direction
-            //   then create the arrow from that and the other data
-            //
-            renderer: (node, child) => {
-                const adaptor = node.adaptor;
-                const {w, h, d} = node.getBBox();
-                const [W, H] = (isVertical ? [h + d, w] : [w, h + d]);
-                const arrow = node.arrow(W, angle, neg, origin, offset, H / 2, double);
-                adaptor.append(node.chtml, arrow);
-            },
-            //
-            // Add the padding to teh proper sides
-            //
-            bbox: Notation.arrowBBox[name],
-            //
-            // Remove redundant notations
-            //
-            remove: remove
-        }] as NotationDefPair;
-    }
-
-};
-
-
-/*****************************************************************/
 /*
  * The CHTMLmenclose wrapper for the MmlMenclose object
  *
@@ -344,33 +55,33 @@ export class CHTMLmenclose<N, T, D> extends CHTMLWrapper<N, T, D> {
             display: 'inline-block',
             left: 0, top: 0,
             position: 'absolute',
-            'border-top': SOLID,
+            'border-top': Notation.SOLID,
             'transform-origin': 'top left'
         },
         'mjx-menclose > mjx-ustrike': {
             display: 'inline-block',
             left: 0, bottom: 0,
             position: 'absolute',
-            'border-top': SOLID,
+            'border-top': Notation.SOLID,
             'transform-origin': 'bottom left'
         },
         'mjx-menclose > mjx-hstrike': {
-            'border-top': SOLID,
+            'border-top': Notation.SOLID,
             position: 'absolute',
             left: 0, right: 0, bottom: '50%',
-            transform: 'translateY(' + (THICKNESS / 2) + 'em)'
+            transform: 'translateY(' + (Notation.THICKNESS / 2) + 'em)'
         },
         'mjx-menclose > mjx-vstrike': {
-            'border-left': SOLID,
+            'border-left': Notation.SOLID,
             position: 'absolute',
             top: 0, bottom: 0, right: '50%',
-            transform: 'translateX(' + (THICKNESS / 2) + 'em)'
+            transform: 'translateX(' + (Notation.THICKNESS / 2) + 'em)'
         },
         'mjx-menclose > mjx-rbox': {
             position: 'absolute',
             top: 0, bottom: 0, right: 0, left: 0,
-            'border': SOLID,
-            'border-radius': (THICKNESS + PADDING) + 'em'
+            'border': Notation.SOLID,
+            'border-radius': (Notation.THICKNESS + Notation.PADDING) + 'em'
         },
         'mjx-menclose > svg.mjx-circle': {
             position: 'absolute',
@@ -378,7 +89,7 @@ export class CHTMLmenclose<N, T, D> extends CHTMLWrapper<N, T, D> {
             width: '100%', height: '100%'
         },
         'mjx-menclose > svg.mjx-circle > ellipse': {
-            'stroke-width': THICKNESS,
+            'stroke-width': Notation.THICKNESS,
             'stroke': 'currentColor',
             'fill': 'none'
         },
@@ -391,7 +102,7 @@ export class CHTMLmenclose<N, T, D> extends CHTMLWrapper<N, T, D> {
         },
         'mjx-menclose > svg.mjx-arrow > line': {
             'stroke': 'currentColor',
-            'stroke-width': THICKNESS,
+            'stroke-width': Notation.THICKNESS,
             'stroke-linecap': 'round'
         },
         'mjx-menclose > svg.mjx-longdiv': {
@@ -401,7 +112,7 @@ export class CHTMLmenclose<N, T, D> extends CHTMLWrapper<N, T, D> {
         },
         'mjx-menclose > svg.mjx-longdiv > path': {
             'stroke': 'currentColor',
-            'stroke-width': THICKNESS,
+            'stroke-width': Notation.THICKNESS,
             'stroke-linecap': 'round',
             'stroke-linejoin': 'round',
             fill: 'none',
@@ -411,7 +122,7 @@ export class CHTMLmenclose<N, T, D> extends CHTMLWrapper<N, T, D> {
     /*
      *  The definitions of the various notations
      */
-    public static Notations: NotationDefList = new Map([
+    public static Notations: Notation.DefList = new Map([
 
         Notation.Border('top'),
         Notation.Border('right'),
@@ -548,13 +259,13 @@ export class CHTMLmenclose<N, T, D> extends CHTMLWrapper<N, T, D> {
             renderChild: true
         }]
 
-    ] as NotationDefPair[]);
+    ] as Notation.DefPair[]);
 
     /*
      *  The notations active on this menclose, and the one to use for the child, if any
      */
-    protected notations: NotationList<N, T, D> = {};
-    protected renderChild: Renderer<N, T, D> = null;
+    protected notations: Notation.List<N, T, D> = {};
+    protected renderChild: Notation.Renderer<N, T, D> = null;
 
     /*
      * fake msqrt for radial notation (if used)
@@ -565,9 +276,9 @@ export class CHTMLmenclose<N, T, D> extends CHTMLWrapper<N, T, D> {
      * The padding, thickness, and shape of the arrow head
      *   (may be overriden using data-padding, data-thickness, and data-arrowhead attibutes)
      */
-    public padding: number = PADDING;
-    public thickness: number = THICKNESS;
-    public arrowhead = {x: ARROWX, y: ARROWY, dx: ARROWDX};
+    public padding: number = Notation.PADDING;
+    public thickness: number = Notation.THICKNESS;
+    public arrowhead = {x: Notation.ARROWX, y: Notation.ARROWY, dx: Notation.ARROWDX};
 
     /*
      * @override
@@ -588,19 +299,19 @@ export class CHTMLmenclose<N, T, D> extends CHTMLWrapper<N, T, D> {
         const attributes = this.node.attributes;
         const padding = attributes.get('data-padding');
         if (padding !== undefined) {
-            this.padding = this.length2em(padding, PADDING);
+            this.padding = this.length2em(padding, Notation.PADDING);
         }
         const thickness = attributes.get('data-thickness');
         if (thickness !== undefined) {
-            this.thickness = this.length2em(thickness, THICKNESS);
+            this.thickness = this.length2em(thickness, Notation.THICKNESS);
         }
         const arrowhead = attributes.get('data-arrowhead') as string;
         if (arrowhead !== undefined) {
             let [x, y, dx] = split(arrowhead);
             this.arrowhead = {
-                x: (x ? parseFloat(x) : ARROWX),
-                y: (y ? parseFloat(y) : ARROWY),
-                dx: (dx ? parseFloat(dx) : ARROWDX)
+                x: (x ? parseFloat(x) : Notation.ARROWX),
+                y: (y ? parseFloat(y) : Notation.ARROWY),
+                dx: (dx ? parseFloat(dx) : Notation.ARROWDX)
             };
         }
     }
@@ -675,8 +386,8 @@ export class CHTMLmenclose<N, T, D> extends CHTMLWrapper<N, T, D> {
         //
         const pbox = this.getPadding();
         const adaptor = this.adaptor;
-        for (const name of Object.keys(sideIndex) as Side[]) {
-            const i = sideIndex[name];
+        for (const name of Notation.sideNames) {
+            const i = Notation.sideIndex[name];
             pbox[i] > 0 && adaptor.setStyle(block, 'padding-' + name, this.em(pbox[i]));
         }
     }
@@ -893,7 +604,7 @@ export class CHTMLmenclose<N, T, D> extends CHTMLWrapper<N, T, D> {
      * @return{N}        The adjusted element
      */
     public adjustThickness(shape: N) {
-        if (this.thickness !== THICKNESS) {
+        if (this.thickness !== Notation.THICKNESS) {
             this.adaptor.setStyle(shape, 'strokeWidth', this.units(this.thickness));
         }
         return shape;
@@ -905,7 +616,7 @@ export class CHTMLmenclose<N, T, D> extends CHTMLWrapper<N, T, D> {
      * @return{N}       The adjusted element
      */
     public adjustBorder(node: N) {
-        if (this.thickness !== THICKNESS) {
+        if (this.thickness !== Notation.THICKNESS) {
             this.adaptor.setStyle(node, 'borderWidth', this.em(this.thickness));
         }
         return node;
@@ -948,4 +659,5 @@ export class CHTMLmenclose<N, T, D> extends CHTMLWrapper<N, T, D> {
         const cbox = this.msqrt.childNodes[0].getBBox();
         return [bbox.h - cbox.h, 0, bbox.d - cbox.d, bbox.w - cbox.w];
     }
+
 }
